requirements.txt
# Core dependencies
torch==2.0.1
numpy>=1.24.3
pandas>=2.0.2
boto3>=1.26.137
scikit-learn>=1.2.2

# AWS and SageMaker specific
sagemaker>=2.132.0
s3fs>=2023.5.0

# Utilities
python-dotenv>=1.0.0
pyarrow>=12.0.1  # For efficient data handling
# torchvision
seaborn
pytz
# logging
# warnings
websocket-client 
polygon-api-client # required for real time execution
alpaca-py 
pymongo
cachetools
optuna